Move provider-specific messages out of background, handle session disconnect
