+++ apps/extension/src/background/index.ts
@@ -7,6 +7,7 @@ import { Ports, KVPrefix } from "router";
 import { chains } from "config";
 import { ChainsService, init as initChains } from "./chains";
 import { KeyRingService, init as initKeyRing } from "./keyring";
+import { SessionHandler } from "./session";
 
 initCrypto();
 initShared();
@@ -22,5 +23,7 @@ const keyRingService = new KeyRingService(store);
 // Initialize messages and handlers
 initChains(router, chainsService);
 initKeyRing(router, keyRingService);
+
+new SessionHandler(keyRingService);
 
 router.listen(Ports.Background);
