+++ apps/extension/src/background/keyring/messages.ts
@@ -12,11 +12,6 @@ enum MessageType {
   GenerateMnemonic = "generate-mnemonic",
   SaveMnemonic = "save-mnemonic",
   DeriveAccount = "derive-account",
-  QueryAccounts = "query-accounts",
-  SignTx = "sign-tx",
-  EncodeTransfer = "encode-transfer",
-  EncodeIbcTransfer = "encode-ibc-transfer",
-  EncodeInitAccount = "encode-init-account",
 }
 
 export class CheckIsLockedMsg extends Message<boolean> {
@@ -203,140 +198,5 @@ export class DeriveAccountMsg extends Message<DerivedAccount> {
 
   type(): string {
     return DeriveAccountMsg.type();
-  }
-}
-
-export class QueryAccountsMsg extends Message<DerivedAccount[]> {
-  public static type(): MessageType {
-    return MessageType.QueryAccounts;
-  }
-
-  constructor(public readonly chainId?: string) {
-    super();
-  }
-
-  validate(): void {
-    return;
-  }
-
-  route(): string {
-    return ROUTE;
-  }
-
-  type(): string {
-    return QueryAccountsMsg.type();
-  }
-}
-
-export class SignTxMsg extends Message<SignedTx> {
-  public static type(): MessageType {
-    return MessageType.SignTx;
-  }
-
-  constructor(
-    public readonly signer: string,
-    public readonly txMsg: string,
-    public readonly txData: string
-  ) {
-    super();
-  }
-
-  validate(): void {
-    if (!this.signer) {
-      throw new Error("A signer address is required to sign transactions!");
-    }
-    if (!this.txMsg || this.txMsg.length === 0) {
-      throw new Error("An encoded txMsg is required!");
-    }
-    if (!this.txData) {
-      throw new Error("txData bytes is required!");
-    }
-    return;
-  }
-
-  route(): string {
-    return ROUTE;
-  }
-
-  type(): string {
-    return SignTxMsg.type();
-  }
-}
-
-export class EncodeTransferMsg extends Message<string> {
-  public static type(): MessageType {
-    return MessageType.EncodeTransfer;
-  }
-
-  constructor(public readonly txMsg: string) {
-    super();
-  }
-
-  validate(): void {
-    if (!this.txMsg) {
-      throw new Error("An encoded txMsg is required!");
-    }
-    return;
-  }
-
-  route(): string {
-    return ROUTE;
-  }
-
-  type(): string {
-    return EncodeTransferMsg.type();
-  }
-}
-
-export class EncodeIbcTransferMsg extends Message<string> {
-  public static type(): MessageType {
-    return MessageType.EncodeIbcTransfer;
-  }
-
-  constructor(public readonly txMsg: string) {
-    super();
-  }
-
-  validate(): void {
-    if (!this.txMsg) {
-      throw new Error("An encoded txMsg is required!");
-    }
-    return;
-  }
-
-  route(): string {
-    return ROUTE;
-  }
-
-  type(): string {
-    return EncodeIbcTransferMsg.type();
-  }
-}
-
-export class EncodeInitAccountMsg extends Message<string> {
-  public static type(): MessageType {
-    return MessageType.EncodeInitAccount;
-  }
-
-  constructor(public readonly txMsg: string, public readonly address: string) {
-    super();
-  }
-
-  validate(): void {
-    if (!this.address) {
-      throw new Error("An address is required!");
-    }
-    if (!this.txMsg) {
-      throw new Error("An encoded txMsg is required!");
-    }
-    return;
-  }
-
-  route(): string {
-    return ROUTE;
-  }
-
-  type(): string {
-    return EncodeInitAccountMsg.type();
   }
 }
