+++ apps/extension/src/provider/index.ts
@@ -1,6 +1,7 @@
 export * from "./Anoma";
 export * from "./InjectedAnoma";
 export * from "./InjectedProxy";
+export * from "./messages";
 export * from "./Proxy";
 export * from "./Signer";
 export * from "./types";
